refactors server code
