+++ server/routes.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const path = require('path');
-const elasticsearch = require('./db/client.js');
+const handlers = require('./handlers.js');
 
 const filePaths = {
     build: path.resolve(__dirname, '../build'),
@@ -19,20 +19,6 @@ module.exports = [
     {
         method: 'POST',
         path: '/delivery',
-        handler: function (request, reply) {
-
-            var data = request.payload;
-            data.userID = request.query.userID;
-            data.timestamp = Date.now();
-
-            // elasticsearch.deleteIndex("ayooo")
-            var result = elasticsearch.addDocument(
-                request.query.type,
-                data
-            ).then( (result) => {
-
-                reply(result);
-            });
-        }
+        handler: handlers.addDeliveryRequest
     }
 ];
